+++ packages/portal-proto/test-utils.tsx
@@ -9,12 +9,21 @@ import {
   DashboardDownloadContext,
 } from "src/utils/contexts";
 import store from "@/app/store";
+import tailwindConfig from "tailwind.config";
 
 const AllTheProviders = ({ children }: { children: React.ReactNode }) => {
+  const themeColors = Object.fromEntries(
+    Object.entries(
+      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+      // @ts-ignore
+      tailwindConfig.plugins.slice(-1)[0].__options.defaultTheme.extend.colors,
+    ).map(([key, values]) => [key, Object.values(values)]),
+  ) as any;
+
   return (
     <CoreProvider>
       <Provider store={store}>
-        <MantineProvider>
+        <MantineProvider theme={{ colors: themeColors }}>
           <URLContext.Provider value={{ prevPath: "", currentPath: "" }}>
             <SummaryModalContext.Provider
               value={{
